fix(geodesic): Dijkstra が 6962 ノード中 1 個しか到達していなかった
`dijkstra` の距離配列が `Float32Array` で、ヒープのキーは倍精度だった。
`dist[e.to] = nd` で丸められ、pop 時に `d > dist[node]` で未丸めの値と比較する。
丸めが下振れするたび（表現できない距離ならほぼ半分）生きたエントリが「古い」と
判定され、その先の部分木ごと捨てられる。

実測：6962 ノードのキャラクタメッシュで、開始オフセット 13mm のとき**到達 1 ノード**。
オフセットぴったり 0 で 91。残りは全部 `Infinity` ＝ `computeAutoWeightsGeodesic`
が「切り離された島」と解釈するので、そのボーンは黙って寄与ゼロになる。
**測地線バインドは今まで機能していなかった。**

既存のユニットテストが全部通っていたのは、距離が 0.5 / 1.5 / 2.5 ―
binary32 で正確に表現できる値ばかりで、一度も丸めが起きないから。

回帰テストは「辺長が表現できない連結グラフで全ノードに到達する」を見る。
旧コードでは `expected 1 to be 800` で落ちる。

+++ src/tools/geodesic-weights.ts
@@ -153,9 +153,28 @@ class MinHeap {
  * Single-source shortest path (Dijkstra) over the mesh graph. `startOffset`
  * seeds the source distance (the Euclidean gap from the bone to the seed node),
  * so geodesic distances are measured from the bone surface, not from zero.
+ *
+ * ## The distance array is Float64Array, and that is load-bearing
+ *
+ * It used to be a `Float32Array`, and the sweep stopped after a handful of
+ * nodes. The heap carries full-precision JS numbers while the array rounds
+ * them, so the staleness check `d > dist[node]` compared an unrounded key
+ * against a rounded store. Whenever the rounding went *down* — about half the
+ * time, for any distance that is not exactly representable — the entry looked
+ * stale and the node was dropped along with everything behind it.
+ *
+ * It is not a small error. On a 6962-node character mesh, seeded at a node with
+ * a 13 mm start offset, the sweep reached **one** node; with an offset of
+ * exactly zero it reached 91. Every other node came back `Infinity`, which
+ * `computeAutoWeightsGeodesic` reads as "this bone is on a disconnected
+ * island" — so the bones silently contributed nothing and the weights fell
+ * through to whichever one or two bones happened to survive.
+ *
+ * The unit test below passed throughout, because its distances are 0.5, 1.5 and
+ * 2.5: exactly representable in binary32, so nothing ever rounded.
  */
-export function dijkstra(graph: MeshGraph, source: number, startOffset: number): Float32Array {
-  const dist = new Float32Array(graph.nodeCount).fill(Infinity);
+export function dijkstra(graph: MeshGraph, source: number, startOffset: number): Float64Array {
+  const dist = new Float64Array(graph.nodeCount).fill(Infinity);
   dist[source] = startOffset;
   const heap = new MinHeap();
   heap.push(source, startOffset);
@@ -198,7 +217,7 @@ export function computeAutoWeightsGeodesic(
   const vertexCount = Math.floor(positions.length / 3);
 
   // Per bone: seed at the node nearest its segment, sweep geodesic distances.
-  const boneGeo: Float32Array[] = [];
+  const boneGeo: Float64Array[] = [];
   const np = new Vector3();
   for (const seg of segments) {
     let bestNode = -1;
@@ -213,7 +232,7 @@ export function computeAutoWeightsGeodesic(
     }
     boneGeo.push(
       bestNode < 0
-        ? new Float32Array(graph.nodeCount).fill(Infinity)
+        ? new Float64Array(graph.nodeCount).fill(Infinity)
         : dijkstra(graph, bestNode, bestD)
     );
   }
